ref: move test cases from SentryHubTests.swift dealing with profiling into new //Tests/SentryTests/Profiling/SentryProfilerTests.swift
